refactor: move view-model interfaces to subdirectory
