+++ src/view-models/interfaces/IPanelBannerViewModel.ts
