+++ src/view-models/interfaces/ISelectionViewModel.ts
